Initialed project

+++ build.sh
@@ -4,8 +4,14 @@ echo "🚀 Running deploy script"
 
 # Reset local changes to avoid git pull conflicts
 echo "[0] 🔄 Resetting local changes"
-git reset --hard HEAD
-git clean -fd
+# Reset only if inside a Git repo
+if [ -d .git ]; then
+  echo "[0] 🔄 Resetting local changes"
+  git reset --hard HEAD
+  git clean -fd
+else
+  echo "⚠️ Not a git repository, skipping reset and clean"
+fi
 
 echo "[1/8] 📥 Pulling latest code from GitHub"
 git pull origin main
